Moved ModelAdapter into sectormodel module
