+++ smif/sectormodel.py
@@ -1,5 +1,25 @@
 from scipy.optimize import minimize
-from smif.abstract import ModelAdapter, ModelInputs
+from smif.abstract import ModelInputs
+
+
+class ModelAdapter(object):
+    """Adapts a model so that it can be used by the optimisation protocol
+
+    Arguments
+    =========
+    model :
+        An instance of a model
+    simulate :
+        The function to use for implementing a `simulate` method
+
+    """
+
+    def __init__(self, model, simulate):
+        self.model = model
+        self.simulate = simulate
+
+    def __getattr__(self, attr):
+        return getattr(self.model, attr)
 
 
 class SectorModel(object):
